backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic[email]==2.10.4
pandas==2.2.3
openpyxl==3.1.5
httpx==0.28.1
slowapi==0.1.9
python-dotenv==1.0.1
